Soportar formato anidado de codes en respuesta de IA
El prompt define que la salida tenga response_id con una lista anidada
codes (assigned_code + comment), pero los DTOs esperaban un Coding plano.
Se agrega AssignedCode, Coding ahora contiene una lista de AssignedCode,
y updateSurveyWithCodedResponses itera la lista anidada generando una
columna por cada code dentro de la respuesta.
